Moving org wide and program event logic out of the helpers file and under Event models
