+++ mysite/helpers.py
@@ -2,17 +2,13 @@ import json
 
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.http import QueryDict
-from django.db.models import Q
-
-from datetime import datetime, timedelta
-from pytz import timezone
-
 from programs.models import Program, Subprogram
+
 
 def paginate_results(request, all_posts):
     page = request.GET.get('page')
     paginator = Paginator(all_posts, 12)
-    
+
     try:
         all_posts = paginator.page(page)
     except PageNotAnInteger:
@@ -25,7 +21,7 @@ def paginate_results(request, all_posts):
 
 def get_org_wide_posts(self, request, page_type, content_model):
     """
-    Function to return a list of content 
+    Function to return a list of content
     for org wide content homepage.
 
     Also checks if there is a query to filter content by initiatives
@@ -42,41 +38,10 @@ def get_org_wide_posts(self, request, page_type, content_model):
         filter_dict['parent_programs'] = int(search_program)
     if date:
         date_range = json.loads(date)
-        filter_dict['date__range'] = (date_range['start'], date_range['end'])   
+        filter_dict['date__range'] = (date_range['start'], date_range['end'])
 
     all_posts = content_model.objects.filter(**filter_dict)
     context['all_posts'] = paginate_results(request, all_posts.live().order_by("-date"))
-    context['programs'] = Program.objects.all().live().in_menu().order_by('title')
-    context['query_url'] = generate_url(request)
-    
-    return context
-
-def get_org_wide_events(self, request, page_type, content_model, tense):
-    """
-    Function to return a list of events 
-    for org wide events homepage.
-
-    Also checks if there is a query to filter content by initiatives
-    or by date for programs.
-    """
-    context = super(page_type, self).get_context(request)
-
-    search_program = request.GET.get('program_id', None)
-    date = request.GET.get('date', None)
-
-    program_query = Q()
-    date_query = set_events_date_query(date, tense)
-
-    if search_program:
-        program_query = Q(parent_programs=int(search_program))
-    
-    all_events = content_model.objects.filter(date_query & program_query)
-
-    if (tense == "future"):
-        context['all_events'] = paginate_results(request, all_events.live().order_by("date", "start_time"))
-    else:
-        context['all_events'] = paginate_results(request, all_events.live().order_by("-date", "-start_time"))
-    
     context['programs'] = Program.objects.all().live().in_menu().order_by('title')
     context['query_url'] = generate_url(request)
 
@@ -85,7 +50,7 @@ def get_org_wide_events(self, request, page_type, content_model, tense):
 
 def get_program_and_subprogram_posts(self, request, page_type, content_model):
     """
-    Function to return a list of content for a program or 
+    Function to return a list of content for a program or
     subprogram content homepage.
 
     Also checks if there is a query to filter content by initiatives
@@ -104,7 +69,7 @@ def get_program_and_subprogram_posts(self, request, page_type, content_model):
         filter_dict = {'parent_programs': program}
         if search_subprogram:
             filter_dict['post_subprogram'] = int(search_subprogram)
-    
+
         all_posts = content_model.objects.live().filter(**filter_dict)
         context['subprograms'] = program.get_children().type(Subprogram).live().order_by('title')
     # if subprogram
@@ -112,7 +77,7 @@ def get_program_and_subprogram_posts(self, request, page_type, content_model):
         subprogram_title = self.get_ancestors()[3]
         program = Subprogram.objects.get(title=subprogram_title)
         all_posts = content_model.objects.live().filter(post_subprogram=program)
-        
+
 
     if date:
         date_range = json.loads(date)
@@ -123,74 +88,6 @@ def get_program_and_subprogram_posts(self, request, page_type, content_model):
     context['program'] = program
 
     return context
-
-def get_program_and_subprogram_events(self, request, page_type, content_model, tense):
-    """
-    Function to return a list of events for a program or 
-    subprogram events homepage.
-
-    Also checks if there is a query to filter content by initiatives
-    or by date for programs.
-    """
-
-    print("in get programs and subprograms!")
-    context = super(page_type, self).get_context(request)
-
-    search_subprogram = request.GET.get('subprogram_id', None)
-    date = request.GET.get('date', None)
-
-    program_query = Q()
-    subprogram_query = Q()
-    date_query = set_events_date_query(date, tense)
-
-    # if program
-    if self.depth == 4:
-        program_title = self.get_ancestors()[2]
-        program = Program.objects.get(title=program_title)
-        program_query = Q(parent_programs=program)
-        if search_subprogram:
-            subprogram_query = Q(post_subprogram=int(search_subprogram))
-
-        context['subprograms'] = program.get_children().type(Subprogram).live().order_by('title')
-    # if subprogram
-    else:
-        subprogram_title = self.get_ancestors()[3]
-        program = Subprogram.objects.get(title=subprogram_title)
-        subprogram_query = Q(post_subprogram=program)
-
-    all_events = content_model.objects.live().filter(date_query & program_query & subprogram_query)
-    
-    if (tense == "future"):
-        context['all_events'] = paginate_results(request, all_events.live().order_by("date", "start_time"))
-    else:
-        context['all_events'] = paginate_results(request, all_events.live().order_by("-date", "-start_time"))
-    context['query_url'] = generate_url(request)
-    context['program'] = program
-    
-    return context
-
-def set_events_date_query(user_date_query, tense):
-    """
-    Function to generate date query for events -
-    If user date query exists, sets this as date range query parameter, 
-    else defaults to greater than current date for future, and less than current date for past
-    """
-    datetime_format = "%Y-%m-%d"
-    eastern = timezone('US/Eastern')
-    curr_date = datetime.now(eastern).date().strftime(datetime_format)
-
-    if user_date_query:
-        date_range = json.loads(user_date_query)
-        date_query = Q(date__range=(date_range['start'], date_range['end']))
-    else:
-        if (tense == "future"):
-            # ensures that multi-day events are displayed on future events page until their end date has passed
-            date_query = Q(date__gte=curr_date) | Q(end_date__gte=curr_date)
-        else:
-            # ensures that multi-day events are not displayed on past events page until their end date has passed
-            date_query = Q(date__lt=curr_date) & (Q(end_date__isnull=True) | Q(end_date__lt=curr_date))
-
-    return date_query
 
 def generate_url(request):
     query = QueryDict(mutable=True)
